samples/14_docker/requirements.txt
google-generativeai>=0.3.0
google-adk>=0.1.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
websockets>=11.0.3 